backend/pyproject.toml
[project]
name = "is-it-open"
version = "0.1.0"
description = ""
authors = [
    {name = "Brandon Lamer-Connolly",email = "brandon.lc.1334@gmail.com"}
]
requires-python = "^3.11"
dependencies = [
    "django (<6.0)",
    "django-ninja (>=1.5.3,<2.0.0)",
    "psycopg[binary] (>=3.3.2,<4.0.0)",
    "gunicorn (>=25.1.0,<26.0.0)",
    "python-decouple (>=3.8,<4.0)",
    "dj-database-url (>=2.1.0,<3.0.0)",
    "requests"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
